+++ Sources/SwiftUIKit/Extension/View+Geometry.swift
@@ -1,0 +1,50 @@
+//
+//  View+Geometry.swift
+//  SwiftUIKit
+//
+//  Created by Daniel Saidi on 2020-03-26.
+//  Copyright © 2020 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+public extension View {
+    
+    /**
+     Bind any `CGFloat` value within a `GeometryProxy` value
+     to an external binding.
+     */
+    func bindGeometry(
+        to value: Binding<CGFloat>,
+        reader: @escaping (GeometryProxy) -> CGFloat) -> some View {
+        self.background(GeometryBinding(reader: reader))
+            .onPreferenceChange(GeometryPreference.self) {
+                value.wrappedValue = $0
+        }
+    }
+}
+
+private struct GeometryBinding: View {
+    
+    let reader: (GeometryProxy) -> CGFloat
+    
+    var body: some View {
+        GeometryReader { geo in
+            Color.clear.preference(
+                key: GeometryPreference.self,
+                value: self.reader(geo)
+            )
+        }
+    }
+}
+
+private struct GeometryPreference: PreferenceKey {
+    
+    typealias Value = CGFloat
+
+    static var defaultValue: CGFloat = 0
+
+    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+        value = max(value, nextValue())
+    }
+}