+++ Sources/SwiftUIKit/Extension/View+SafeArea.swift
@@ -11,7 +11,7 @@ import SwiftUI
 public extension View {
     
     /**
-     Read the safe area insets of a certain edge and bind it
+     Bind the safe area insets of a certain edge and bind it
      to the provided binding parameter.
      
      This modifier is very useful when you want a view to be
@@ -19,21 +19,18 @@ public extension View {
      the previously ignored safe areas. Just use the binding
      to set the edge padding of the view you want to inset.
      */
-    func readSafeAreaInsets(
-        edge: Edge,
-        binding: Binding<CGFloat>) -> some View {
-        self.background(InsetGetter(edge: edge))
-            .onPreferenceChange(InsetPreferenceKey.self) {
-                binding.wrappedValue = $0
+    func bindSafeAreaInsets(
+        for edge: Edge,
+        to binding: Binding<CGFloat>) -> some View {
+        self.bindGeometry(to: binding) {
+            self.inset(for: $0, edge: edge)
         }
     }
 }
 
-private struct InsetGetter: View {
+private extension View {
     
-    let edge: Edge
-    
-    func inset(for geo: GeometryProxy) -> CGFloat {
+    func inset(for geo: GeometryProxy, edge: Edge) -> CGFloat {
         let insets = geo.safeAreaInsets
         switch edge {
         case .top: return insets.top
@@ -41,25 +38,5 @@ private struct InsetGetter: View {
         case .leading: return insets.leading
         case .trailing: return insets.trailing
         }
-    }
-    
-    var body: some View {
-        GeometryReader { geometry in
-            Rectangle().preference(
-                key: InsetPreferenceKey.self,
-                value: self.inset(for: geometry)
-            )
-        }
-    }
-}
-
-private struct InsetPreferenceKey: PreferenceKey {
-    
-    typealias Value = CGFloat
-
-    static var defaultValue: CGFloat = 0
-
-    static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
-        value = max(value, nextValue())
     }
 }